+++ NEWS.md
@@ -1,4 +1,24 @@
+# burnr v0.2.0.9000
+
+* Added 'site_stats()' for site-level sumary statistics akin to FHX2 and FHAES
+
+* Updated composite to include a filter for number of events. This necesitates fixes in other functions (e.g. plot_demograph) 
+
+* Updated run_sea with new enhanced speed, and changed the "time_span" parameter to "key_period" for a simpler input
+
 # burnr v0.1.2.9000
+
+# burnr v0.1.1
+
+Changes in this patch:
+
+* Added option to sort series by their last (or most recent) year with `sort()`.
+
+* Updated `run_sea()` to better replicate EVENT Fortran program. Addresses #58. Added detailed description and an example page with graphics. Also fixed issue with percentage CIs in the departure table. Mind you, `run_sea()` is still super experimental.
+
+* Several minor changes to documentation and meta data.
+
+# burnr v0.1.0
 
 Changes in this minor release: